Got rid of unnecessary functions that merely duplicated other functions, Instead refactoredexperience gain function to pass back an object under specified parameters.
